+++ checkpoint/CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.11.20] - 2025-07-15
 
 ### Changed
 
 - #v1 Add LeafHandlerRegistry and make it as default for V1 codes
+- ThreadedCheckpointDeleter with a new background_thread_count parameter so that
+  users can adjust how many threads to handle old checkpoint deletion.
 
 ## [0.11.19] - 2025-07-08